requirements.txt
arrow==1.2.2
asgiref==3.5.0
binaryornot==0.4.4
certifi==2022.6.15
chardet==5.0.0
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.5
commonmark==0.9.1
cookiecutter==1.6.0
crispy-tailwind==0.5.0
defusedxml==0.7.1
Django==4.0.3
django-browser-reload==1.6.0
django-crispy-forms==1.14.0
django-environ==0.9.0
django-tailwind==3.3.0
djangorestframework==3.13.1
environ==1.0
future==0.18.2
httpie==3.2.1
idna==3.3
Jinja2==3.1.2
jinja2-time==0.2.0
MarkupSafe==2.1.1
multidict==6.0.2
Pillow==9.0.1
poyo==0.5.0
psycopg2==2.9.3
psycopg2-binary==2.9.3
Pygments==2.12.0
PySocks==1.7.1
python-dateutil==2.8.2
python-slugify==6.1.2
pytz==2022.1
PyYAML==6.0
requests==2.28.1
requests-toolbelt==0.9.1
rich==12.5.1
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
tzdata==2021.5
urllib3==1.26.10
whichcraft==0.6.1
whitenoise==6.2.0
